Fix slow '5' in zsh.
The key for kill-word was bound to the wrong sequence.
Blame https://unix.stackexchange.com/a/583783.
It only worked because the real escape sequences
contains it. But, if you just typed 5, it would wait KEYTIMEOUT
for you to complete the sequence, causing notable lag when typing
the painfully common number 5.